% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_table.R
\docType{data}
\name{icon_table}
\alias{icon_table}
\title{icon_table}
\format{
A data frame with 1,538 rows and 5 variables:
\describe{
\item{\code{name}}{character. Name of the icon.}
\item{\code{style}}{character. Style of the icon.}
\item{\code{full_name}}{character. Full name of the icon.}
\item{\code{svg}}{character. SVG output text.}
\item{\code{terms}}{character. Terms for categorizing the icons.}
}
}
\usage{
icon_table
}
\description{
A table of icons generated using the internal \code{fontawesome:::fa_tbl} from the
\code{fontawesome} R package by Rstudio.
}
\keyword{datasets}
