% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myicons-package.R
\docType{package}
\name{myicons-package}
\alias{myicons}
\alias{myicons-package}
\title{myicons: myicons}
\description{
Manage your own personal icon library and generate
    random icons for quick prototyping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/myicons}
  \item Report bugs at \url{https://github.com/jimbrig/myicons/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
