% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_icon.R
\name{random_icon}
\alias{random_icon}
\title{Random Icon Generator}
\usage{
random_icon(term = NULL, class = NULL)
}
\arguments{
\item{term}{(optional) term to limit number of icons available to a specific
category. For a list of available terms see the \link{icon_table}.}

\item{class}{Additional classes to customize the style of the icon (see the
\href{http://fontawesome.io/examples/}{usage examples} for details on
supported styles).}
}
\value{
An HTML icon element returned via \code{\link[shiny:icon]{shiny::icon()}}
}
\description{
This function generates a random \code{fontawesome} icon for quick usage
while prototyping shiny apps.
}
\examples{
require(shiny)
random_icon() # should display in R Viewer
random_icon("animal")
shiny::fluidRow(random_icon())
}
